requirements.txt
streamlit
matplotlib
numpy
scipy
biopython
pysam
pandas
ont-remora
pod5
ruptures